feat: Adiciona TAG para nomear mais amigávelmente na documentação Swagger nos controllers. Além de em Livrosimplementa findByStatusAndUsuarioId

+++ src/main/java/br/com/dsacramento/booklist/service/LivroService.java
@@ -2,11 +2,14 @@ package br.com.dsacramento.booklist.service;
 
 import br.com.dsacramento.booklist.entity.Categoria;
 import br.com.dsacramento.booklist.entity.Livro;
-import br.com.dsacramento.booklist.entity.Pessoa;
+import br.com.dsacramento.booklist.entity.Usuario;
+import br.com.dsacramento.booklist.enums.StatusLivro;
 import br.com.dsacramento.booklist.repository.CategoriaRepository;
 import br.com.dsacramento.booklist.repository.LivroRepository;
-import br.com.dsacramento.booklist.repository.PessoaRepository;
+import br.com.dsacramento.booklist.repository.UsuarioRepository;
+import br.com.dsacramento.booklist.security.CustomUserDetails;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -19,32 +22,38 @@ import java.util.Set;
 public class LivroService {
 
     private final LivroRepository livroRepository;
-    private final PessoaRepository pessoaRepository;
+    private final UsuarioRepository usuarioRepository;
     private final CategoriaRepository categoriaRepository;
 
     public List<Livro> findAll() {
-        return livroRepository.findAll();
+        return livroRepository.findByUsuarioId(getUsuarioLogado().getId());
     }
 
     public Livro findById(Long id) {
-        return livroRepository.findById(id)
+        Livro livro = livroRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Livro não encontrado!"));
+
+        if (!livro.getUsuario().getId().equals(getUsuarioLogado().getId())){
+            throw new RuntimeException("Acesso negado a este livro!");
+        }
+        return livro;
+    }
+
+    public List<Livro> findByStatus(StatusLivro statusLivro) {
+        Long usuarioId = getUsuarioLogado().getId();
+
+        return livroRepository.findByStatusLivroAndUsuarioId(statusLivro, usuarioId);
     }
 
     @Transactional
-    public Livro save(Livro livro) {
+    public Livro save(Livro livro, Long usuarioId) {
 
-        if (livro.getPessoa() != null && livro.getPessoa().getId() != null) {
-            Pessoa pessoa = pessoaRepository.findById(livro.getPessoa().getId())
-                    .orElseThrow(() -> new RuntimeException("Pessoa não encontrada com o ID informado!"));
+        Usuario usuarioLogado = usuarioRepository.findById(usuarioId)
+                .orElseThrow(() -> new RuntimeException("Usuário não encontrado!"));
 
-            livro.setPessoa(pessoa);
-        } else {
-            throw new RuntimeException("O livro precisa estar obrigatoriamente vinculado a uma pessoa!");
-        }
+        livro.setUsuario(usuarioLogado);
 
         if (livro.getCategorias() != null && !livro.getCategorias().isEmpty()) {
-
             Set<Categoria> categorias = new HashSet<>();
 
             for (Categoria categoriaRequest : livro.getCategorias()) {
@@ -64,6 +73,19 @@ public class LivroService {
     @Transactional
     public void delete(Long id) {
         Livro livro = findById(id);
+
+        CustomUserDetails userDetails =(CustomUserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if (!livro.getUsuario().getId().equals(userDetails.getId())){
+            throw new RuntimeException("Você não tem permmissão para deletar este livro!");
+        }
+
         livroRepository.delete(livro);
+    }
+
+    private CustomUserDetails getUsuarioLogado() {
+        return (CustomUserDetails) SecurityContextHolder.getContext()
+                .getAuthentication()
+                .getPrincipal();
     }
 }
